Fixed an out of bounds error.
Usage of `size_t` was causing a loop condition to always be met as the loop iterator could
never have the sentinal value -1 and thus the loop could never correctly exit. Usage
`size_t` of also caused index iterator to wrap to max value of `size_t` when initialized
with -1, which occured when there were zero local variables.

+++ include/compiler.h
@@ -8,6 +8,7 @@
 #define clox_compiler_h
 
 #include "chunk.h"
+#include "common.h"
 #include "object.h"
 #include "scanner.h"
 
@@ -43,7 +44,7 @@ typedef enum {
  */
 typedef struct {
     Token name;
-    size_t depth;
+    intmax_t depth;
 } Local;
 
 /**
@@ -51,8 +52,8 @@ typedef struct {
  */
 typedef struct {
     Local locals[UINT8_COUNT];
-    size_t localCount;
-    size_t scopeDepth;
+    intmax_t localCount;
+    intmax_t scopeDepth;
 } Compiler;
 
 /**
